+++ src/main/kotlin/com/mituuz/fuzzier/Fuzzier.kt
@@ -89,8 +89,8 @@ open class Fuzzier : FuzzyAction() {
         }
     }
 
-    override fun createPopup(): JBPopup {
-        val popup = getInitialPopup()
+    override fun createPopup(screenDimensionKey: String): JBPopup {
+        val popup = getInitialPopup(screenDimensionKey)
 
         popup.addListener(object : JBPopupListener {
             override fun onClosed(event: LightweightWindowEvent) {