+++ src/main/kotlin/com/mituuz/fuzzier/settings/FuzzierGlobalSettingsConfigurable.kt
@@ -43,6 +43,8 @@ class FuzzierGlobalSettingsConfigurable : Configurable {
         component = FuzzierGlobalSettingsComponent()
         component.newTabSelect.getCheckBox().isSelected = state.newTab
         component.recentFileModeSelector.getRecentFilesTypeComboBox().selectedIndex = state.recentFilesMode.ordinal
+        component.defaultHeight.getIntSpinner().value = state.defaultPopupHeight
+        component.defaultWidth.getIntSpinner().value = state.defaultPopupWidth
         component.searchPosition.getSearchPositionComboBox().selectedIndex = state.searchPosition.ordinal
         component.prioritizeShortDirs.getCheckBox().isSelected = state.prioritizeShorterDirPaths
         component.debounceTimerValue.getIntSpinner().value = state.debouncePeriod
@@ -67,6 +69,8 @@ class FuzzierGlobalSettingsConfigurable : Configurable {
     override fun isModified(): Boolean {
         return state.newTab != component.newTabSelect.getCheckBox().isSelected
                 || state.recentFilesMode != component.recentFileModeSelector.getRecentFilesTypeComboBox().selectedItem
+                || state.defaultPopupHeight != component.defaultHeight.getIntSpinner().value
+                || state.defaultPopupWidth != component.defaultWidth.getIntSpinner().value
                 || state.searchPosition != component.searchPosition.getSearchPositionComboBox().selectedItem
                 || state.prioritizeShorterDirPaths != component.prioritizeShortDirs.getCheckBox().isSelected
                 || state.debouncePeriod != component.debounceTimerValue.getIntSpinner().value
@@ -90,13 +94,20 @@ class FuzzierGlobalSettingsConfigurable : Configurable {
         state.newTab = component.newTabSelect.getCheckBox().isSelected
         state.recentFilesMode = RecentFilesMode.entries.toTypedArray()[component.recentFileModeSelector.getRecentFilesTypeComboBox().selectedIndex]
 
-        val selectedSearchPosition: FuzzierGlobalSettingsService.SearchPosition = FuzzierGlobalSettingsService.SearchPosition.entries.toTypedArray()[component.searchPosition.getSearchPositionComboBox().selectedIndex]
-        if (state.searchPosition != selectedSearchPosition) {
+        val newPopupHeight = component.defaultHeight.getIntSpinner().value as Int
+        val newPopupWidth = component.defaultWidth.getIntSpinner().value as Int
+        val newSearchPosition: FuzzierGlobalSettingsService.SearchPosition = FuzzierGlobalSettingsService.SearchPosition.entries.toTypedArray()[component.searchPosition.getSearchPositionComboBox().selectedIndex]
+        if (state.searchPosition != newSearchPosition ||
+            state.defaultPopupHeight != newPopupHeight ||
+            state.defaultPopupWidth != newPopupWidth) {
+
             // Reset window size and split position to defaults
             state.resetWindow = true
             state.splitPosition = FuzzierGlobalSettingsService.DEFAULT_SPLIT_POSITION
-            state.searchPosition = selectedSearchPosition
         }
+        state.defaultPopupHeight = newPopupHeight
+        state.defaultPopupWidth = newPopupWidth
+        state.searchPosition = newSearchPosition
         state.prioritizeShorterDirPaths = component.prioritizeShortDirs.getCheckBox().isSelected
         state.debouncePeriod = component.debounceTimerValue.getIntSpinner().value as Int
         state.fileListLimit = component.fileListLimit.getIntSpinner().value as Int